py3.9 improvement
