+++ preset_build.py
@@ -31,7 +31,6 @@ if __name__ == '__main__':
         "mcwzh-meme_legacy_noresource_sfw.zip",
         "mcwzh-meme.zip"
     ]
-    pack_builder = build.builder()
     pack_counter = 0
     perfect_pack_counter = 0
     base_folder = "builds"
@@ -41,17 +40,16 @@ if __name__ == '__main__':
         os.mkdir(base_folder)
     for file in os.listdir(base_folder):
         os.remove(os.path.join(base_folder, file))
-    for item, name in zip(preset_args, preset_name):
-        pack_builder.args = item
-        pack_builder.build()
-        if pack_builder.error_count == 0:
+    for args, name in zip(preset_args, preset_name):
+        info, warning_count, error_count = build.build(args)
+        if error_count == 0:
             pack_counter += 1
-            if pack_builder.warning_count == 0:
+            if warning_count == 0:
                 perfect_pack_counter += 1
             if name != "mcwzh-meme.zip":
                 os.rename("builds/mcwzh-meme.zip",
                           os.path.join(base_folder, name))
-            print(f"Renamed pack to {name}.")
+                print(f"Renamed pack to {name}.")
         else:
             print(f"Failed to build pack {name}.")
     print(
